Engineering Task #3461 (1 hours): Adjusted data type of id from primitive to object. Added toString methods to all domain classes.
git-svn-id: http://mcoltrac/svn/mcol-app/sdt/trunk@17531 e0420d7e-bc29-8447-b031-85aca2f896c5

+++ dao/src/integ-test/java/uk/gov/moj/sdt/dao/BulkSubmissionDaoTest.java
@@ -167,7 +167,7 @@ public class BulkSubmissionDaoTest extends AbstractTransactionalJUnit4SpringCont
 
         for (IIndividualRequest req : submission.getIndividualRequests ())
         {
-            Assert.assertEquals (5, req.getId ());
+            Assert.assertEquals (5, req.getId ().longValue ());
         }
 
     }
